Remove separate validate method, change path to an array
The aim of this change is twofold:

Simplify the API by removing the validate method, which just wrapped
decode and added contextual the path parameter to it (i.e. where we are
currently in the object tree).  After this change, the path is
calculated on-demand in non-leaf combinators like `array`, `object` or
`record` when we encounter an error. Leaf codecs like `string` or
`number` always return an empty array as the path while the spine codecs
fill in the full path for each error.

Improve the performance by not having to calculate the path in the happy
case.

Currently the path manipulation in spine codecs is done by mutating the
path array, but let's see if we want to change that to immutable code at
some point.

Thanks to @milankinen for the idea